Huge refactoring
