+++ src/app.js
@@ -1,45 +1,33 @@
 import React from "react";
-import { Route, NavLink, Switch, Redirect } from "react-router-dom";
+import { Route, Switch, withRouter } from "react-router-dom";
 import "antd/dist/antd.css";
-import OrderForm from "./components/order-form";
 import Cart from "./components/cart";
-import Filter from "./components/filter";
 import RestaurantsPage from "./components/routes/restaurants";
-import CheckoutPage from "./components/routes/checkout";
+import CheckoutPage from "./components/routes/cart-checkout";
 import { Provider } from "./contexts/username";
-import { useInputValue } from "./custom-hooks/use-input-value";
-import { Input } from "antd";
 import Menu, { MenuItem } from "./components/menu";
 
-export default function App() {
-  const [username, setUserName] = useInputValue("Roma");
-
+function App() {
   return (
     <div>
       <Menu>
-        <Menu.Item to="/checkout">
+        <MenuItem to="/checkout">
           <Cart />
-        </Menu.Item>
+        </MenuItem>
         <MenuItem to="/restaurants">Restaurants</MenuItem>
-        <MenuItem to="/filter" children={"Filter"} />
       </Menu>
-      <div>
-        Username: <Input value={username} onChange={setUserName} />
-      </div>
-      <Provider value={username}>
-        <Switch>
-          <Redirect from="/" exact to="/restaurants" />
-          <Redirect from="/restaurants/" exact strict to="/restaurants" />
-          <Route path="/filter" exact component={Filter} />
-          <Route path="/checkout" exact component={CheckoutPage} />
-          <Route
-            path="/restaurants/:id/review"
-            render={({ id }) => <h1>Add a review for {id}</h1>}
-          />
-          <Route path="/restaurants" component={RestaurantsPage} />
-          <Route path="*" render={() => <h1>Not Found Page</h1>} />
-        </Switch>
-      </Provider>
+      <Switch>
+        <Route path="/" exact render={() => null} />
+        <Route path="/checkout" exact component={CheckoutPage} />
+        <Route path="/restaurants" component={RestaurantsPage} />
+        <Route
+          path="/restaurants/:id/review"
+          render={({ id }) => <h1>Add a review for {id}</h1>}
+        />
+        <Route path="*" render={() => <h1>Not Found Page</h1>} />
+      </Switch>
     </div>
   );
 }
+
+export default withRouter(App);
