+++ src/components/dish.js
@@ -1,11 +1,12 @@
 import React from "react";
-import { Card, Button } from "antd";
+import { NavLink } from "react-router-dom";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { addItem, removeItem } from "../ac";
-import { dishSelector } from "../selectors";
+import { Card, Button } from "antd";
+import { addItem, removeItem } from "../reducer/order/actions";
+import { dishRestaurantSelector, dishSelector } from "../selectors";
 
-function Dish({ dish, amount, handleDecrease, handleIncrease }) {
+function Dish({ dish, amount, handleDecrease, handleIncrease, restaurant }) {
   return (
     <Card
       bordered
@@ -34,7 +35,12 @@ function Dish({ dish, amount, handleDecrease, handleIncrease }) {
         </>
       ]}
     >
-      <Card.Meta title={dish.name} description={dish.ingredients.join(", ")} />
+      <NavLink to={`/restaurants/${restaurant.id}`}>
+        <Card.Meta
+          title={dish.name}
+          description={dish.ingredients.join(", ")}
+        />
+      </NavLink>
     </Card>
   );
 }
@@ -44,14 +50,21 @@ Dish.defaultProps = {
 };
 
 Dish.propTypes = {
-  price: PropTypes.number,
-  name: PropTypes.string,
-  ingredients: PropTypes.arrayOf(PropTypes.string)
+  dish: PropTypes.shape({
+    id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
+    name: PropTypes.string,
+    price: PropTypes.number,
+    ingredients: PropTypes.arrayOf(PropTypes.string)
+  }),
+  amount: PropTypes.number,
+  handleDecrease: PropTypes.func,
+  handleIncrease: PropTypes.func
 };
 
 const mapStateToProps = (state, ownProps) => ({
   amount: state.order.get(ownProps.id) || 0,
-  dish: dishSelector(state, ownProps)
+  dish: dishSelector(state, ownProps),
+  restaurant: dishRestaurantSelector(state, ownProps)
 });
 
 const mapDispatchToProps = {
