+++ src/components/restaurant.js
@@ -1,11 +1,11 @@
 import React from "react";
-import { Avatar, Button, List } from "antd";
+import { connect } from "react-redux";
 import PropTypes from "prop-types";
+import { Avatar, List } from "antd";
 import ReviewList from "./review-list";
 import RestaurantMenu from "./restaurant-menu";
 import RestaurantMap from "./restaurant-map";
 import RestaurantRate from "./restaurant-rate";
-import { connect } from "react-redux";
 import { restaurantSelector } from "../selectors";
 
 function Restaurant({ restaurant, isOpen }) {
