--- src/components/routes/checkout.js
@@ -1,10 +1,0 @@
-import React from "react";
-import Checkout from "../checkout";
-
-function CheckoutPage() {
-  return <Checkout />;
-}
-
-CheckoutPage.propTypes = {};
-
-export default CheckoutPage;
