--- src/constants/index.js
@@ -1,18 +1,0 @@
-export const INCREMENT = "INCREMENT";
-export const DECREMENT = "DECREMENT";
-
-export const ADD_ITEM = "ADD_ITEM";
-export const REMOVE_ITEM = "REMOVE_ITEM";
-
-export const SET_MIN_RATING = "SET_MIN_RATING";
-
-export const ADD_REVIEW = "ADD_REVIEW";
-export const LOAD_ALL_REVIEWS = "LOAD_ALL_REVIEWS";
-
-export const LOAD_ALL_RESTAURANTS = "LOAD_ALL_RESTAURANTS";
-
-export const LOAD_MENU = "LOAD_MENU";
-
-export const START = "_START";
-export const SUCCESS = "_SUCCESS";
-export const ERROR = "_ERROR";
