+++ src/contexts/username.js
@@ -1,6 +1,6 @@
 import { createContext } from "react";
 
-export const userContext = createContext("unknown user");
+export const userContext = createContext("Anonymous user");
 const { Provider, Consumer } = userContext;
 
 export { Provider, Consumer };
