+++ src/reducer/dishes/constants.js
@@ -1,0 +1,1 @@
+export const LOAD_MENU = "LOAD_MENU";
