+++ src/reducer/filters/constants.js
@@ -1,0 +1,1 @@
+export const SET_MIN_RATING = "SET_MIN_RATING";
