+++ src/reducer/restaurants/constants.js
@@ -1,0 +1,5 @@
+export const LOAD_ALL_RESTAURANTS = "LOAD_ALL_RESTAURANTS";
+
+export const START = "_START";
+export const SUCCESS = "_SUCCESS";
+export const ERROR = "_ERROR";
