+++ src/reducer/reviews/constants.js
@@ -1,0 +1,2 @@
+export const ADD_REVIEW = "ADD_REVIEW";
+export const LOAD_ALL_REVIEWS = "LOAD_ALL_REVIEWS";
