+++ src/store/index.js
@@ -2,7 +2,6 @@ import { createStore, applyMiddleware } from "redux";
 import thunk from "redux-thunk";
 import { routerMiddleware } from "connected-react-router";
 import reducer from "../reducer";
-import logger from "../middlewares/logger";
 import generateId from "../middlewares/generateId";
 import api from "../middlewares/api";
 import history from "../history";
